creacion de xml formulario para activity_perfil.xml, strings.xml, correcion de otros archivos
