+++ Aplicaciones de Moviles/app/app/src/main/java/com/example/filmcast/ui/login/BuscarActivity.kt
@@ -5,11 +5,11 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import com.example.filmcast.R
 
-class ActorActivity : AppCompatActivity() {
+class BuscarActivity : AppCompatActivity() {
 
-    override fun onCreate(savedInstanceState: Bundle?) {
+    override fun onCreate(savedInstanceState: Bundle?){
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_actor)
+        setContentView(R.layout.activity_buscar)
     }
 }
