+++ Aplicaciones de Moviles/app/app/src/main/java/com/example/filmcast/ui/login/MainActivity.kt
@@ -23,7 +23,7 @@ class MainActivity : AppCompatActivity() {
             val intent = Intent(this, menuActivity::class.java)
             startActivity(intent)
 
-            //--ydfgydg-->
+            //--ydfggtytsfydg-->
         }
     }
 }
