+++ Aplicaciones de Moviles/app/app/src/main/java/com/example/filmcast/ui/login/PerfilActivity.kt
@@ -3,13 +3,12 @@ package com.example.filmcast.ui.login
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import com.example.filmcast.R
 
-class ActorActivity : AppCompatActivity() {
+class PerfilActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
+    super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_actor)
+        setContentView(R.layout.activity_perfil)
     }
 }
