--- Aplicaciones de Moviles/app/app/src/main/res/values-land/dimens.xml
@@ -1,4 +1,0 @@
-<resources>
-    <dimen name="fab_margin">48dp</dimen>
-    <dimen name="activity_horizontal_margin">48dp</dimen>
-</resources>
