--- Aplicaciones de Moviles/app/app/src/main/res/values-w1240dp/dimens.xml
@@ -1,4 +1,0 @@
-<resources>
-    <dimen name="fab_margin">200dp</dimen>
-    <dimen name="activity_horizontal_margin">200dp</dimen>
-</resources>
