--- Aplicaciones de Moviles/app/app/src/main/res/values/dimens.xml
@@ -1,6 +1,0 @@
-<resources>
-    <dimen name="fab_margin">16dp</dimen>
-    <!-- Default screen margins, per the Android Design guidelines. -->
-    <dimen name="activity_horizontal_margin">16dp</dimen>
-    <dimen name="activity_vertical_margin">16dp</dimen>
-</resources>
