+++ Aplicaciones de Moviles/app/app/src/main/res/values/strings.xml
@@ -124,6 +124,7 @@
     <string name="Pe_genero">Género</string>
     <string name="Pe_rol">Rol</string>
     <string name="Pe_cerrar_sesion">Cerrar sesión</string>
+    <string name="Pe_editar_perfil">Editar Perfil</string>
 
 
     <string-array name="Bu_edad">
@@ -241,51 +242,28 @@
         <item>24.001 o más</item>
     </string-array>
 
-
-    <string name="lorem_ipsum">
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam in scelerisque sem. Mauris
-        volutpat, dolor id interdum ullamcorper, risus dolor egestas lectus, sit amet mattis purus
-        dui nec risus. Maecenas non sodales nisi, vel dictum dolor. Class aptent taciti sociosqu ad
-        litora torquent per conubia nostra, per inceptos himenaeos. Suspendisse blandit eleifend
-        diam, vel rutrum tellus vulputate quis. Aliquam eget libero aliquet, imperdiet nisl a,
-        ornare ex. Sed rhoncus est ut libero porta lobortis. Fusce in dictum tellus.\n\n
-        Suspendisse interdum ornare ante. Aliquam nec cursus lorem. Morbi id magna felis. Vivamus
-        egestas, est a condimentum egestas, turpis nisl iaculis ipsum, in dictum tellus dolor sed
-        neque. Morbi tellus erat, dapibus ut sem a, iaculis tincidunt dui. Interdum et malesuada
-        fames ac ante ipsum primis in faucibus. Curabitur et eros porttitor, ultricies urna vitae,
-        molestie nibh. Phasellus at commodo eros, non aliquet metus. Sed maximus nisl nec dolor
-        bibendum, vel congue leo egestas.\n\n
-        Sed interdum tortor nibh, in sagittis risus mollis quis. Curabitur mi odio, condimentum sit
-        amet auctor at, mollis non turpis. Nullam pretium libero vestibulum, finibus orci vel,
-        molestie quam. Fusce blandit tincidunt nulla, quis sollicitudin libero facilisis et. Integer
-        interdum nunc ligula, et fermentum metus hendrerit id. Vestibulum lectus felis, dictum at
-        lacinia sit amet, tristique id quam. Cras eu consequat dui. Suspendisse sodales nunc ligula,
-        in lobortis sem porta sed. Integer id ultrices magna, in luctus elit. Sed a pellentesque
-        est.\n\n
-        Aenean nunc velit, lacinia sed dolor sed, ultrices viverra nulla. Etiam a venenatis nibh.
-        Morbi laoreet, tortor sed facilisis varius, nibh orci rhoncus nulla, id elementum leo dui
-        non lorem. Nam mollis ipsum quis auctor varius. Quisque elementum eu libero sed commodo. In
-        eros nisl, imperdiet vel imperdiet et, scelerisque a mauris. Pellentesque varius ex nunc,
-        quis imperdiet eros placerat ac. Duis finibus orci et est auctor tincidunt. Sed non viverra
-        ipsum. Nunc quis augue egestas, cursus lorem at, molestie sem. Morbi a consectetur ipsum, a
-        placerat diam. Etiam vulputate dignissim convallis. Integer faucibus mauris sit amet finibus
-        convallis.\n\n
-        Phasellus in aliquet mi. Pellentesque habitant morbi tristique senectus et netus et
-        malesuada fames ac turpis egestas. In volutpat arcu ut felis sagittis, in finibus massa
-        gravida. Pellentesque id tellus orci. Integer dictum, lorem sed efficitur ullamcorper,
-        libero justo consectetur ipsum, in mollis nisl ex sed nisl. Donec maximus ullamcorper
-        sodales. Praesent bibendum rhoncus tellus nec feugiat. In a ornare nulla. Donec rhoncus
-        libero vel nunc consequat, quis tincidunt nisl eleifend. Cras bibendum enim a justo luctus
-        vestibulum. Fusce dictum libero quis erat maximus, vitae volutpat diam dignissim.
-    </string>
+    //registro
     <string name="tv_registro">Registrarse</string>
     <string name="hint_username">Nombre de usuario</string>
     <string name="tv_login">Iniciar Sesion</string>
     <string name="hint_password">Contraseña</string>
+    <string name="hint_confpassword">Confirmar Contraseña</string>
     <string name="btn_register">Registrarse</string>
     <string name="tv_or">Or</string>
     <string name="tv_terms_conditions">By continuing you are agreed the terms and conditions and privacy policy.</string>
     <string name="tv_terms_conditions1">Al continuar, usted acepta los términos y condiciones, así como la política de privacidad.</string>
     <string name="hint_email">email</string>
     <string name="btn_login">Iniciar Sesion</string>
+    
+    
+    //editar perfil 
+    <string name="EDP_Titulo">Editar Perfil</string>
+    <string name="EDP_Nombre">Nombre de usuario</string>
+    <string name="EDP_Descripcion">Descripcion</string>
+    <string name="EDP_Telefono">Telefono</string>
+    <string name="EDP_Nacionalidad">Nacionalidad</string>
+    <string name="EDP_Email">Email</string>
+    <string name="EDP_cancelar">Cancelar</string>
+    <string name="EDP_Guardar">Guardar</string>
+    
 </resources>
